% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeCpp11.R
\docType{package}
\name{EncodeCpp11-package}
\alias{EncodeCpp11}
\alias{EncodeCpp11-package}
\title{EncodeCpp11: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
